[SEDONA 710] Rename Geostats SQL classes to generic name; merge UdfRegistrator into AbstractCatalog
