+++ spark/common/src/main/scala/org/apache/spark/sql/sedona_sql/plans/logical/EvalPhysicalFunction.scala
@@ -16,22 +16,24 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.spark.sql.sedona_sql.strategy.geostats
+package org.apache.spark.sql.sedona_sql.plans.logical
 
-import org.apache.spark.sql.Strategy
+import org.apache.spark.sql.catalyst.expressions.Attribute
+import org.apache.spark.sql.catalyst.expressions.AttributeSet
+import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
-import org.apache.spark.sql.execution.SparkPlan
-import org.apache.spark.sql.sedona_sql.plans.logical.EvalGeoStatsFunction
-import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.sedona_sql.expressions.ST_GeoStatsFunction
+import org.apache.spark.sql.catalyst.plans.logical.UnaryNode
 
-class EvalGeoStatsFunctionStrategy(spark: SparkSession) extends Strategy {
+case class EvalPhysicalFunction(
+    function: Expression,
+    resultAttrs: Seq[Attribute],
+    child: LogicalPlan)
+    extends UnaryNode {
 
-  override def apply(plan: LogicalPlan): Seq[SparkPlan] = {
-    plan match {
-      case EvalGeoStatsFunction(function: ST_GeoStatsFunction, resultAttrs, child) =>
-        EvalGeoStatsFunctionExec(function, planLater(child), resultAttrs) :: Nil
-      case _ => Nil
-    }
-  }
+  override def output: Seq[Attribute] = child.output ++ resultAttrs
+
+  override def producedAttributes: AttributeSet = AttributeSet(resultAttrs)
+
+  override protected def withNewChildInternal(newChild: LogicalPlan): LogicalPlan =
+    copy(child = newChild)
 }
